+++ core/src/main/java/org/sterl/spring/persistent_tasks/task/exception/FailTaskNoRetryException.java
@@ -5,6 +5,7 @@ package org.sterl.spring.persistent_tasks.task.exception;
  */
 public class FailTaskNoRetryException extends TaskException {
     private static final long serialVersionUID = 1L;
+
     public FailTaskNoRetryException(String message, Throwable cause) {
         super(message, cause);
     }